+++ src-core/src/main.rs
@@ -80,6 +80,8 @@ fn configure_command_handlers() -> impl Fn(tauri::Invoke) {
         openvr::commands::openvr_set_fade_distance,
         openvr::commands::openvr_set_image_brightness,
         openvr::commands::openvr_launch_binding_configuration,
+        openvr::commands::openvr_get_binding_origins,
+        openvr::commands::openvr_is_dashboard_visible,
         os::commands::run_command,
         os::commands::play_sound,
         os::commands::show_in_folder,