+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Updated translations to use ICU syntax.
+- Migrated from Legacy OpenVR Input system to SteamVR's current input system (Controller bindings are now configured in SteamVR!)
  
 ### Fixed
 - Fixed failed installation of dotnet hosting bundle preventing retry installation of individual runtimes.